Add capability to run init scripts iteratively
This allows iteratively executing init scripts in the /docker-entrypoints/ folder of the container. It sets the basis for services which will require multiple init scripts.

This will be the case, for example, when we will need to create services for development, which will require additional inits, for example we will need to run npm ci in the v3 service which already runs merge_json.sh.